Organizing namespaces for new files.
