+++ src/ForeverFactory/Generators/Transforms/Factories/BaseRecursiveTransformFactory.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using System.Reflection;
+using ForeverFactory.Generators.Transforms.Factories.ReflectionTargets;
 
 namespace ForeverFactory.Generators.Transforms.Factories
 {
@@ -70,51 +71,6 @@ namespace ForeverFactory.Generators.Transforms.Factories
         private bool CanApplyRecursion(TargetInfo targetInfo)
         {
             return _options.EnableRecursiveInstantiation && targetInfo.TargetType != typeof(string);
-        }
-    }
-    
-    internal abstract class TargetInfo
-    {
-        public abstract Type TargetType { get; }
-        public abstract string Name { get; }
-            
-        public abstract void SetValue(object instance, object value);
-    };
-
-    internal sealed class PropertyTargetInfo : TargetInfo
-    {
-        private readonly PropertyInfo _propertyInfo;
-
-        public override Type TargetType => _propertyInfo.PropertyType;
-        public override string Name => _propertyInfo.Name;
-            
-
-        public PropertyTargetInfo(PropertyInfo propertyInfo)
-        {
-            _propertyInfo = propertyInfo;
-        }
-
-        public override void SetValue(object instance, object value)
-        {
-            _propertyInfo.SetValue(instance, value);
-        }
-    }
-        
-    internal sealed class FieldTargetInfo : TargetInfo
-    {
-        private readonly FieldInfo _fieldInfo;
-            
-        public override Type TargetType => _fieldInfo.FieldType;
-        public override string Name => _fieldInfo.Name;
-
-        public FieldTargetInfo(FieldInfo fieldInfo)
-        {
-            _fieldInfo = fieldInfo;
-        }
-
-        public override void SetValue(object instance, object value)
-        {
-            _fieldInfo.SetValue(instance, value);
         }
     }
 }
